% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTable.R
\name{cleanTable}
\alias{cleanTable}
\title{Parsing text string to a table}
\usage{
cleanTable(txt, rowsplit = "|", newline = "\\\\n")
}
\arguments{
\item{txt}{Single text string from sendPrompt to attempt to parse into a table}

\item{rowsplit}{Character used to split rows of the table}

\item{newline}{Character used to identify a new line}
}
\value{
A dataframe parsed from the text string
}
\description{
This function takes a text string from the sendPrompt function and attempts to parse it to a dataframe. Much of the parsing was done via trial and error in trying to understand how Gemini returns tables. In general, it is important to specify the column names in the sendPrompt function and how the text should be delimited. The default values in this function should work ok when no particular instructions were given, but it is very important to verify the output.
}
\examples{
\dontrun{
cleanTable(sendPrompt(prompt="Return a table of the 5 largest US states by population. Include the state name and the total population.", type="text", apiKey="###"))
}
}
