% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countTokens.R
\name{countTokens}
\alias{countTokens}
\title{Count number of tokens in a prompt}
\usage{
countTokens(
  prompt,
  type = "text",
  apiKey = NULL,
  image = NULL,
  model = "gemini-1.5-flash"
)
}
\arguments{
\item{prompt}{Text prompt to send to Gemini. Should be a single text string.}

\item{type}{Can be either "text" or "image". If "image" the directory location or url of the image should be given.}

\item{apiKey}{Google Developer's API key. Needed to run the function. See https://ai.google.dev/gemini-api/docs/api-key}

\item{image}{If type is "image" this should be the local directory location of the image or a url.}

\item{model}{Gemini model variant to be used. Default is "gemini-1.5-flash" which can be used for text or images. See https://ai.google.dev/gemini-api/docs/models/gemini for other options.}
}
\value{
Number of tokens used in the prompt.
}
\description{
Count number of tokens in a prompt
}
\examples{
\dontrun{
#Simple text  prompt
countTokens(prompt="Hi, how are you?", apiKey="###")
}
}
